+++ packages/terra-core-docs/src/terra-dev-site/doc/signature/example/SignatureExample.jsx
@@ -1,6 +1,10 @@
 import React from 'react';
 import Signature from 'terra-signature';
 import classNames from 'classnames/bind';
+import Tabs from 'terra-tabs';
+import VisuallyHiddenText from 'terra-visually-hidden-text';
+import TextSignature from './TextSignature';
+import ImageSignature from './ImageSignature';
 import styles from './SignatureExample.module.scss';
 
 const cx = classNames.bind(styles);
@@ -9,73 +13,73 @@ class SignatureExample extends React.Component {
   constructor() {
     super();
 
-    this.state = { lineSegments: [], lineWidth: Signature.Opts.Width.FINE };
-
-    this.handleSingleLine = this.handleSingleLine.bind(this);
+    this.state = { lineSegments: [], lineWidth: Signature.Opts.Width.FINE, message: '' };
     this.handleClear = this.handleClear.bind(this);
     this.handleLineWidth = this.handleLineWidth.bind(this);
-  }
-
-  handleSingleLine() {
-    const singleLine = [{
-      x1: 71, y1: 8, x2: 71, y2: 8,
-    }, {
-      x1: 71, y1: 8, x2: 71, y2: 10,
-    }, {
-      x1: 71, y1: 10, x2: 71, y2: 17,
-    }, {
-      x1: 71, y1: 17, x2: 71, y2: 28,
-    }, {
-      x1: 71, y1: 28, x2: 71, y2: 44,
-    }, {
-      x1: 71, y1: 44, x2: 71, y2: 56,
-    }, {
-      x1: 71, y1: 56, x2: 71, y2: 68,
-    }, {
-      x1: 71, y1: 68, x2: 71, y2: 75,
-    }, {
-      x1: 71, y1: 75, x2: 71, y2: 81,
-    }, {
-      x1: 71, y1: 81, x2: 71, y2: 84,
-    }, {
-      x1: 71, y1: 84, x2: 71, y2: 86,
-    }, {
-      x1: 71, y1: 86, x2: 71, y2: 87,
-    }, {
-      x1: 71, y1: 87, x2: 71, y2: 87,
-    }];
-    const newState = { ...this.state, lineSegments: singleLine };
-    this.setState(newState);
+    this.handleLiveRegion = this.handleLiveRegion.bind(this);
+    this.tabKey = 'compact';
   }
 
   handleClear() {
     const newState = { ...this.state, lineSegments: [] };
     this.setState(newState);
+    this.handleLiveRegion();
   }
 
   handleLineWidth(event) {
     this.setState({ [event.target.name]: event.target.value });
   }
 
+  handleLiveRegion(value) {
+    this.setState({ message: value ? 'Added Signature' : 'Cleared Signature' });
+    setTimeout(() => {
+      this.setState({ message: '' });
+    }, 3000);
+  }
+
   render() {
     return (
       <div>
-        <div className={cx('signature-wrapper')}>
-          <Signature id="foo" lineWidth={this.state.lineWidth} lineSegments={this.state.lineSegments} />
-        </div>
-        <div>
-          <button type="button" onClick={this.handleClear}>Clear </button>
-          <button type="button" onClick={this.handleSingleLine}>Sign w/Line </button>
-          <div>
-            <p><label htmlFor="lineWidth">Select a line width:</label></p>
-            <select id="lineWidth" name="lineWidth" value={this.state.lineWidth} onChange={this.handleLineWidth}>
-              <option value="1">EXTRAFINE</option>
-              <option value="2">FINE</option>
-              <option value="4">MEDIUM</option>
-              <option value="6">HEAVY</option>
-            </select>
-          </div>
-        </div>
+        <Tabs setFocusOnContent defaultActiveKey={`${this.tabKey}DrawTab`} id={this.tabKey}>
+          <Tabs.Pane label="Draw" key={`${this.tabKey}DrawTab`} id={`${this.tabKey}DrawTab`}>
+            <br />
+            <>
+              <div className={cx('signature-wrapper')}>
+                <Signature id="draw" lineWidth={this.state.lineWidth} lineSegments={this.state.lineSegments} />
+              </div>
+            </>
+            <br />
+            <div>
+              <button type="button" onClick={this.handleClear}>Clear </button>
+              <div>
+                <p><label htmlFor="lineWidth">Select a line width:</label></p>
+                <select id="lineWidth" name="lineWidth" value={this.state.lineWidth} onChange={this.handleLineWidth}>
+                  <option value="1">EXTRAFINE</option>
+                  <option value="2">FINE</option>
+                  <option value="4">MEDIUM</option>
+                  <option value="6">HEAVY</option>
+                </select>
+              </div>
+            </div>
+          </Tabs.Pane>
+          <Tabs.Pane label="Type" key={`${this.tabKey}TextTab`} id={`${this.tabKey}TextTab`}>
+            <br />
+            <TextSignature onClearSignature={this.handleLiveRegion} onAddSignature={this.handleLiveRegion} />
+          </Tabs.Pane>
+          <Tabs.Pane label="Upload" key={`${this.tabKey}UploadTab`} id={`${this.tabKey}UploadTab`}>
+            <br />
+            <ImageSignature onClearSignature={this.handleLiveRegion} />
+          </Tabs.Pane>
+        </Tabs>
+        <p>
+          <strong>Note: </strong>
+          <br />
+          <ul>
+            <li>This example Terra Signature currently supports only Draw Functionality.</li>
+            <li>Signature Text and Signature Upload Image functionalities will be part of future enhancement.</li>
+          </ul>
+        </p>
+        <VisuallyHiddenText aria-live="polite" text={this.state.message} />
       </div>
     );
   }